# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "description": "Erro no formato RFC 9457 (application/problem+json)",
        "properties": {
          "code": {
            "description": "Código estável para o front (ex.: NAO_ENCONTRADO)",
            "example": "TRANSICAO_INVALIDA",
            "type": "string"
          },
          "detail": {
            "description": "Ausente em erros 5xx",
            "example": "Uma solicitação com status Aberta não pode ser aprovada. Inicie a análise primeiro.",
            "type": "string"
          },
          "errors": {
            "description": "Erros por campo (DADOS_INVALIDOS)",
            "example": [
              {
                "campo": "titulo",
                "mensagem": "O título deve ter pelo menos 5 caracteres."
              }
            ],
            "items": {
              "properties": {
                "campo": {
                  "description": "Caminho do campo, com \".\" nos aninhados (ex.: item.quantidade)",
                  "type": "string"
                },
                "mensagem": {
                  "type": "string"
                }
              },
              "required": [
                "campo",
                "mensagem"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "instance": {
            "description": "Caminho da requisição",
            "example": "/api/v1/solicitacoes/0b6c8f9e-3d2a-4f7b-9c1e-5a8d2f4b6c10/decisao",
            "type": "string"
          },
          "requestId": {
            "description": "Mesmo valor do header X-Request-Id",
            "example": "f3c1a2b4-6d8e-4f0a-9b1c-3e5d7f9a2b46",
            "type": "string"
          },
          "status": {
            "example": 409,
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "example": "Transição de status inválida",
            "type": "string"
          },
          "type": {
            "example": "https://solicitacoes.local/erros/transicao-invalida",
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "instance",
          "code"
        ],
        "type": "object"
      },
...
      "UsuarioAtualDto": {
        "properties": {
          "area": {
            "example": {
              "id": "1a3c5e7b-9d2f-4b46-8c0e-7f9a1b3d5e28",
              "nome": "Financeiro"
            },
            "properties": {
              "id": {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              "nome": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "nome"
            ],
            "type": "object"
          },
          "cargo": {
            "enum": [
              "SOLICITANTE",
              "ANALISTA",
              "ADMIN"
            ],
            "example": "SOLICITANTE",
            "type": "string"
          },
          "email": {
            "example": "ana.souza@demo.test",
            "type": "string"
          },
          "id": {
            "example": "3e7b1c9a-2d4f-4a68-8b0e-6c1d3f5a7b92",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "nome": {
            "example": "Ana Souza",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "cargo",
          "area"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "API para registro, análise e acompanhamento de solicitações internas.",
    "title": "Gestão de Solicitações Internas",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "usuarioAtual",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioAtualDto"
                }
              }
            },
            "description": "Usuário atual"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sem token, token inválido ou sessão encerrada (NAO_AUTENTICADO)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Usuário autenticado, com a área",
        "tags": [
          "autenticação"
        ]
      }
    },
...
    "/api/v1/solicitacoes": {
      "get": {
        "operationId": "listarSolicitacoes",
        "parameters": [
          {
            "description": "Título, descrição ou código (SOL-000042 ou 42), sem diferenciar maiúsculas e acentos. Depois de tirar os espaços nas pontas: vazio (ignorado) ou de 2 a 200 caracteres",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ABERTA",
                    "EM_ANALISE",
                    "APROVADA",
                    "REJEITADA"
                  ],
                  "type": "string"
                },
                {
                  "items": {
                    "enum": [
                      "ABERTA",
                      "EM_ANALISE",
                      "APROVADA",
                      "REJEITADA"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "prioridade",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "BAIXA",
                    "MEDIA",
                    "ALTA"
                  ],
                  "type": "string"
                },
                {
                  "items": {
                    "enum": [
                      "BAIXA",
                      "MEDIA",
                      "ALTA"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                }
              ]
            }
          },
          {
            "description": "prioridade: ALTA → MEDIA → BAIXA e, dentro, a mais antiga primeiro (ignora direcao)",
            "in": "query",
            "name": "ordenarPor",
            "required": false,
            "schema": {
              "default": "dataSolicitacao",
              "enum": [
                "dataSolicitacao",
                "prioridade"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direcao",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "description": "eu: só as que têm o usuário atual como analista responsável",
            "in": "query",
            "name": "analista",
            "required": false,
            "schema": {
              "enum": [
                "eu"
              ],
              "type": "string"
            }
          },
          {
            "description": "Id da área (UUID); pode repetir: area=<id>&area=<id>",
            "in": "query",
            "name": "area",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                },
                {
                  "items": {
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                    "type": "string"
                  },
                  "type": "array"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 9007199254740991,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaSolicitacoesDto"
                }
              }
            },
            "description": "Página de solicitações"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Parâmetros inválidos (DADOS_INVALIDOS)"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sem token, token inválido ou sessão encerrada (NAO_AUTENTICADO)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lista com busca, filtros, ordenação e paginação (escopo do usuário)",
        "tags": [
          "solicitações"
        ]
      },
      "post": {
        "operationId": "criarSolicitacao",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarSolicitacaoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitacaoDto"
                }
              }
            },
            "description": "Criada; o header Location aponta para ela",
            "headers": {
              "Location": {
                "description": "/api/v1/solicitacoes/{id}",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Corpo inválido ou com campos não aceitos (DADOS_INVALIDOS)"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sem token, token inválido ou sessão encerrada (NAO_AUTENTICADO)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Abre uma solicitação (nasce ABERTA)",
        "tags": [
          "solicitações"
        ]
      }
    },
    "/api/v1/solicitacoes/{id}": {
      "delete": {
        "operationId": "excluirSolicitacao",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Excluída"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sem token, token inválido ou sessão encerrada (NAO_AUTENTICADO)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Não é o dono nem administrador (ACESSO_NEGADO)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Não existe, foi excluída ou não é visível para o usuário (SOLICITACAO_NAO_ENCONTRADA)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Status não permite exclusão (EDICAO_BLOQUEADA)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Exclusão lógica (dono com ABERTA ou administrador sem decisão)",
        "tags": [
          "solicitações"
        ]
      },
      "get": {
        "operationId": "detalharSolicitacao",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitacaoDto"
                }
              }
            },
            "description": "Solicitação"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sem token, token inválido ou sessão encerrada (NAO_AUTENTICADO)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Não existe, foi excluída ou não é visível para o usuário (SOLICITACAO_NAO_ENCONTRADA)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalhe, com as ações permitidas ao usuário atual",
        "tags": [
          "solicitações"
        ]
      },
      "patch": {
        "operationId": "editarSolicitacao",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditarSolicitacaoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitacaoDto"
                }
              }
            },
            "description": "Solicitação atualizada"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Corpo inválido ou com campos não aceitos (DADOS_INVALIDOS)"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sem token, token inválido ou sessão encerrada (NAO_AUTENTICADO)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Não é o dono nem administrador (ACESSO_NEGADO)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Não existe, foi excluída ou não é visível para o usuário (SOLICITACAO_NAO_ENCONTRADA)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Status não permite edição (EDICAO_BLOQUEADA) ou versão desatualizada (CONFLITO_DE_VERSAO)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edita título, descrição e prioridade (dono com ABERTA ou administrador sem decisão)",
        "tags": [
          "solicitações"
        ]
      }
    },
    "/api/v1/solicitacoes/{id}/analise": {
      "post": {
        "operationId": "iniciarAnalise",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitacaoDto"
                }
              }
            },
            "description": "Solicitação em análise"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sem token, token inválido ou sessão encerrada (NAO_AUTENTICADO)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Cargo sem permissão (ACESSO_NEGADO) ou a própria solicitação (SEGREGACAO_DE_FUNCOES)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Não existe, foi excluída ou não é visível para o usuário (SOLICITACAO_NAO_ENCONTRADA)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A solicitação não está ABERTA (TRANSICAO_INVALIDA)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Inicia a análise (ABERTA → EM_ANALISE); quem inicia vira o responsável",
        "tags": [
          "solicitações"
        ]
      }
    },
    "/api/v1/solicitacoes/{id}/decisao": {
      "post": {
        "operationId": "decidirSolicitacao",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisaoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitacaoDto"
                }
              }
            },
            "description": "Solicitação decidida"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Corpo inválido ou com campos não aceitos (DADOS_INVALIDOS)"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sem token, token inválido ou sessão encerrada (NAO_AUTENTICADO)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Cargo sem permissão ou não é o responsável (ACESSO_NEGADO), ou a própria solicitação (SEGREGACAO_DE_FUNCOES)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Não existe, foi excluída ou não é visível para o usuário (SOLICITACAO_NAO_ENCONTRADA)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A solicitação não está EM_ANALISE (TRANSICAO_INVALIDA)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aprova ou rejeita com comentário (analista responsável ou administrador)",
        "tags": [
          "solicitações"
        ]
      }
    },
    "/api/v1/solicitacoes/{id}/historico": {
      "get": {
        "operationId": "historicoSolicitacao",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventoHistoricoDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Eventos"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sem token, token inválido ou sessão encerrada (NAO_AUTENTICADO)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Não existe, foi excluída ou não é visível para o usuário (SOLICITACAO_NAO_ENCONTRADA)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Linha do tempo, do evento mais antigo para o mais recente",
        "tags": [
          "solicitações"
        ]
      }
    },
...
  },
  "servers": [],
...
}